refactor: split cluster commands into cluster/zone subcommands
