+++ implementations/rust/ockam/ockam_command/src/lib.rs
@@ -84,3 +84,4 @@ pub mod value_parsers;
 mod vault;
 mod version;
 mod worker;
+mod zone;
