+++ implementations/rust/ockam/ockam_command/src/zone/static/repl/long_about.txt
